+++ client/lib/ft4/admin/admin-operations.ts
@@ -4,6 +4,7 @@ import { authDescriptor as authDesc } from "../accounts/auth-descriptor";
 import { AuthDescriptor } from "../accounts/auth-descriptor/types";
 import { BufferId } from "/cryptoUtils";
 import { Amount } from "../asset/interfaces";
+import { Asset } from "../asset/types";
 
 export function registerAccount(authDescriptor: AuthDescriptor): Operation {
   const ad = authDesc.toGtv(authDescriptor);
@@ -12,12 +13,12 @@ export function registerAccount(authDescriptor: AuthDescriptor): Operation {
 
 export function addRateLimitPoints(
   accountId: BufferId,
-  amount: number
+  amount: number,
 ): Operation {
   return op(
     "ft4.admin.add_rate_limit_points",
     formatter.ensureBuffer(accountId),
-    amount
+    amount,
   );
 }
 
@@ -25,7 +26,7 @@ export function registerAsset(
   name: string,
   symbol: string,
   decimals: number,
-  iconUrl: string
+  iconUrl: string,
 ): Operation {
   return op("ft4.admin.register_asset", name, symbol, decimals, iconUrl);
 }
@@ -33,12 +34,27 @@ export function registerAsset(
 export function mint(
   accountId: BufferId,
   assetId: BufferId,
-  amount: Amount
+  amount: Amount,
 ): Operation {
   return op(
     "ft4.admin.mint",
     formatter.ensureBuffer(accountId),
     formatter.ensureBuffer(assetId),
-    amount.value
+    amount.value,
+  );
+}
+
+export function registerCrosschainAsset(
+  asset: Asset,
+  originBrid: BufferId,
+): Operation {
+  return op(
+    "ft4.admin.register_crosschain_asset",
+    asset.name,
+    asset.symbol,
+    asset.decimals,
+    asset.brid,
+    asset.iconUrl,
+    formatter.ensureBuffer(originBrid),
   );
 }