+++ client/lib/ft4/admin/index.ts
@@ -1,0 +1,7 @@
+export {
+  registerAccount,
+  addRateLimitPoints,
+  registerAsset,
+  mint,
+  registerCrosschainAsset,
+} from "./admin-op-functions";