+++ client/lib/ft4/asset/asset-query-functions.ts
@@ -17,18 +17,16 @@ export async function getAssetById(
   connection: Connection,
   id: BufferId,
 ): Promise<Asset> {
-  return await connection
-    .query<AssetResponse>(assetById(id))
-    .then(createAssetObject);
+  const response = await connection.query(assetById(id));
+  return response ? createAssetObject(response) : null;
 }
 
 export async function getAssetBySymbol(
   connection: Connection,
   symbol: string,
 ): Promise<Asset> {
-  return await connection
-    .query<AssetResponse>(assetBySymbol(symbol))
-    .then(createAssetObject);
+  const response = await connection.query(assetBySymbol(symbol));
+  return response ? createAssetObject(response) : null;
 }
 
 export function getAssetsByName(
@@ -63,7 +61,7 @@ export async function getBalanceByAccountId(
   assetId: BufferId,
 ): Promise<Balance> {
   return await connection
-    .query<BalanceResponse>(balanceByAccountId(accountId, assetId))
+    .query(balanceByAccountId(accountId, assetId))
     .then(createBalanceObject);
 }
 
@@ -71,9 +69,7 @@ export async function getBalancesByAccountId(
   connection: Connection,
   accountId: BufferId,
 ): Promise<Balance[]> {
-  const balances = await connection.query<BalanceResponse[]>(
-    balancesByAccountId(accountId),
-  );
+  const balances = await connection.query(balancesByAccountId(accountId));
   return balances.map(createBalanceObject);
 }