+++ client/lib/ft4/authentication/types.ts
@@ -10,7 +10,6 @@ export interface Authenticator {
   // TODO: check if authDataService can be removed
   authDataService: AuthDataService;
   createSession(): AuthenticatorSession;
-  getAuthFlags(operation: Operation): Promise<string[]>;
   getKeyHandlerForOperation(
     operation: Operation,
   ): Promise<KeyHandler | undefined>;
@@ -59,7 +58,7 @@ export interface AuthDataService {
     accountId: BufferId,
     authDescriptorId: BufferId,
   ): Promise<number | null>;
-  getLoginConfig(name: string | null): Promise<LoginConfig>;
+  getLoginConfig(name: string | undefined): Promise<LoginConfig>;
   getBrid(): Buffer;
 }