+++ client/lib/ft4/utils/index.ts
@@ -1,24 +1,24 @@
 import {
-  Operation as newOperation,
+  Operation,
   encryption,
-  QueryArguments,
   RawGtv,
   IClient,
+  gtv,
+  RawGtx,
 } from "postchain-client";
 import { Config } from "./types";
+import { Buffer } from "buffer";
 
-export function nop(): newOperation {
+export function nop(): Operation {
   return { name: "nop", args: [encryption.randomBytes(32)] };
 }
 
-export function op(name: string, ...args: readonly RawGtv[]): newOperation {
+export function op(name: string, ...args: readonly RawGtv[]): Operation {
   return { name, args: args as RawGtv[] };
 }
 
 export async function getConfig(session: IClient): Promise<Config> {
-  const response = await session.query<QueryArguments, ConfigResponse>(
-    "ft4.get_config"
-  );
+  const response = await session.query<ConfigResponse>("ft4.get_config");
   return Object.freeze({
     rateLimit: {
       active: response.rate_limit.active,
@@ -29,10 +29,12 @@ export async function getConfig(session: IClient): Promise<Config> {
   });
 }
 
+export function getTransactionRid(tx: RawGtx): Buffer {
+  return gtv.gtvHash(tx[0]); //tx body
+}
+
 export async function getVersion(session: IClient): Promise<string> {
-  return Object.freeze(
-    await session.query<QueryArguments, string>("ft4.get_version")
-  );
+  return Object.freeze(await session.query<string>("ft4.get_version"));
 }
 
 type DeepReadonly<T> = T extends (infer R)[]