+++ client/lib/ft4/utils/types.ts
@@ -1,5 +1,5 @@
-import { RellOperation } from "postchain-client";
-import { OptionalPageCursor } from "../types";
+import { RellOperation, TransactionReceipt } from "postchain-client";
+import { OptionalPageCursor, Session } from "../types";
 import { Buffer } from "buffer";
 
 export type Config = {
@@ -45,14 +45,14 @@ export type PaginatedEntity<T> = {
 };
 
 export type TxBuilderTransaction = {
-  blockchainRID: Buffer;
+  blockchainRid: Buffer;
   operations: RellOperation[];
   signers: Buffer[];
   signatures: Buffer[];
 };
 
 export interface RellAppStructure {
-  modules: Record<string, RellModuleStructure>;
+  [modules: string]: Record<string, RellModuleStructure>;
 }
 
 export interface RellModuleStructure {
@@ -63,3 +63,14 @@ export interface RellOperationStructure {
   mount: string;
   parameters: any[];
 }
+
+export type RequireTogether<T, Keys extends keyof T> = T & {
+  [K in Keys]-?: T[K];
+};
+export type TransactionCompletion<T = undefined> = T extends undefined
+  ? { receipt: TransactionReceipt }
+  : { receipt: TransactionReceipt; data: T };
+
+export type TransactionSessionCompletion<T = undefined> = T extends undefined
+  ? { receipt: TransactionReceipt; session: Session }
+  : { receipt: TransactionReceipt; session: Session; data: T };