+++ changelog.md
@@ -5,17 +5,37 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.1.5] -- 2023-09-12
+## [0.1.7] - 2023-10-19
+
+### Added
+- Added createOrchestrator for crosschain transfers
+- Refactored type exports to allow easier importing from entry index file.
+- assetOriginById: query that retrieves the "asset origin", which is the only chain the asset can be received from
+- findPathToChainForAsset: traverses the tree structure of the linked chains to find the path to a certain asset.
+- `TransactionBuilder` now has a function `buildAndSend` which immediately submits the built transaction
+- Functions that add operations to `TransactionBuilder` now accepts an optional callback which will be invoked when the transaction is included in a block that has been anchored on the anchoring chain
+
+
+## [0.1.6] - 2023-09-29
+
+### Fixed
+- addAuthDescriptor and deleteAuthDescriptor were hard to use, as you couldn't easily use the new keypair you just added in subsequent operations. They now return the receipt and a new session to use for future operations if you want to also use the current auth descriptor.
+- exported some types regarding assets that weren't available for end users
+
+### Changed
+- All operations now return an TransactionCompletion, which holds the receipt and (optionally) additional data
+
+## [0.1.5] - 2023-09-12
 
 ### Added
 - createGenericEvmKeyStore: it receives an address and a sign function, to allow for custom implementations with any web3 library. Metamask is still supported through ethers for ease of setup.
 - fixed examples
 - Custom Event Emitter for handling various events like Metamask address change, crosschain transfer notifications.
-- External signing messages now includes rid of the blockchain to which the tx is being submitted.
+- Auth messages now includes rid of the blockchain to which the tx is being submitted.
 
 ### Changed
 - Transfer history's asset properties are now of the Asset type
-- Blockchain rid is now part of signing message when using external signatures
+- Auth messages now include rid of the blockchain to which the tx is being submitted.
 
 ### Fixed
 - Asset queries now return Asset type with `iconUrl`, not `icon_url`